Pre-production version (but fully working)
This setup is for P12 heads and microphones. For use with Happy or the original capsule design, remove the two 680R SMD resistors outlined with a dotted mark and insert 2x 2K2 resistors as shown on the render. Although this works with the P12 Varee, I've done a simpler version to support that type of setup.

A version for the THAT1512 will be available later (it's done but needs to go to production/testing)